Add Color Legend behavior
Now the color legend bar is no longer static. It will be hidden if it isn't necessary: when dimensions to render are less than 3 and there isn't any rotation belonging to the fourth dimension or higher. When dimension of rotations or mesh dimensions are higher than 4 the coordinate of the label will be updated.