Add "flatCell" param to MeshND
Now "MeshND" class has a new parameter called "flatCells" and it is used to store the flat cells, which are the two-dimensional faces of a hyperobject. "#createHypercubeFlatCells()" is responsible for creating all the square faces present in a hypercube (e.g. 6 square faces in a cube and 24 square faces in a tesseract).